pyproject.toml
[tool.poetry]
name = "llm-anonymizer"
version = "0.1.0"
description = "Python package to anonymize documents"
authors = ["Prince Ngattai Lam <ngattai.lam@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.10.3"
python-dotenv = "1.0.0"
litellm = "^1.55.3"
faker = "^33.1.0"
pip = "^24.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
